# frontend/src/openapi_v5.yaml
openapi: 3.0.3
info:
  title: Empire Trading Hub - Unified API Gateway
  description: |
    Enterprise-grade multi-broker trading platform API.
    Supports TradeLocker, Topstep (Topstep), and TruForex (MT4/MT5).
    
    **Authentication**: Bearer token (API key from broker account registration)
    **Rate Limiting**: Plan-based (Starter: 100/min, Pro: 500/min, Elite: 2000/min)
    **Webhooks**: HMAC-SHA256 signed payloads
  version: 5.0.0
  contact:
    name: Empire Trading Support
    email: support@empiretrading.io
  license:
    name: Proprietary
servers:
  - url: https://api.empiretrading.io/api/unify/v1
    description: Production
  - url: https://staging-api.empiretrading.io/api/unify/v1
    description: Staging
security:
  - BearerAuth: []
...
paths:
...
  /positions:
    get:
      tags: [Positions]
      summary: Get all open positions
      operationId: getPositions
      parameters:
        - name: broker
          in: query
          schema:
            type: string
            enum: [tradelocker, topstep, truforex]
        - name: symbol
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
  /positions/{positionId}:
    put:
      tags: [Positions]
      summary: Modify position SL/TP
      operationId: modifyPosition
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stop_loss:
                  type: number
                  format: double
                take_profit:
                  type: number
                  format: double
      responses:
        '200':
          description: Position modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
    delete:
      tags: [Positions]
      summary: Close a position
      operationId: closePosition
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Position closed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        API key generated during broker account registration.
        Include in Authorization header: `Bearer your-api-key`
    HmacAuth:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: |
        HMAC-SHA256 signature of webhook payload.
        Format: `t=<timestamp>,v1=<signature>`
  schemas:
...
    Position:
      type: object
      properties:
        id:
          type: string
          format: uuid
        broker:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [LONG, SHORT]
        qty:
          type: number
          format: double
        entry_price:
          type: number
          format: double
        current_price:
          type: number
          format: double
        unrealized_pnl:
          type: number
          format: double
        unrealized_pnl_pct:
          type: number
          format: double
        stop_loss:
          type: number
          format: double
        take_profit:
          type: number
          format: double
        open_time:
          type: string
          format: date-time
...